Handle directories not existing

+++ index.js
@@ -3,6 +3,7 @@ const cp = require('child_process');
 const os = require('os');
 const path = require('path');
 const crypto = require('crypto');
+const fs = require('fs');
 
 /*::
 type OptionsType = {
@@ -14,6 +15,9 @@ type OptionsType = {
 module.exports = function fastRMRF(directory /*: string */, options /*: OptionsType */ = {}) {
 	const tmpDir = options.tmpDir || os.tmpdir();
 	const tmpDirPath = path.join(tmpDir, getHash());
+  if (!fs.existsSync(directory)) {
+    return null;
+  }
   cp.execSync(`mkdir -p ${tmpDirPath}`);
   cp.execSync(`mv ${directory} ${tmpDirPath}`);
   const child = cp.spawn('rm', ['-rf', tmpDirPath], {
